improved disable/delete with relationships

+++ app/Http/Controllers/Settings/Currencies.php
@@ -95,9 +95,20 @@ class Currencies extends Controller
      */
     public function update(Currency $currency, Request $request)
     {
-        $canDisable = $currency->canDisable();
+        $relationships = $this->countRelationships($currency, [
+            'accounts' => 'accounts',
+            'customers' => 'customers',
+            'invoices' => 'invoices',
+            'revenues' => 'revenues',
+            'bills' => 'bills',
+            'payments' => 'payments',
+        ]);
 
-        if ($canDisable === true || $request['enabled']) {
+        if ($currency->code == setting('general.default_currency')) {
+            $relationships[] = strtolower(trans_choice('general.companies', 1));
+        }
+
+        if (empty($relationships) || $request['enabled']) {
             $currency->update($request->all());
 
             // Update default currency setting
@@ -109,43 +120,15 @@ class Currencies extends Controller
             $message = trans('messages.success.updated', ['type' => trans_choice('general.currencies', 1)]);
 
             flash($message)->success();
+
+            return redirect('settings/currencies');
         } else {
-            $text = array();
-
-            if (isset($canDisable['company'])) {
-                $text[] = '<b>' . $canDisable['company'] . '</b> ' . trans_choice('general.companies', ($canDisable['company'] > 1) ? 2 : 1);
-            }
-
-            if (isset($canDisable['accounts'])) {
-                $text[] = '<b>' . $canDisable['accounts'] . '</b> ' . trans_choice('general.accounts', ($canDisable['accounts'] > 1) ? 2 : 1);
-            }
-
-            if (isset($canDisable['customers'])) {
-                $text[] = '<b>' . $canDisable['customers'] . '</b> ' . trans_choice('general.customers', ($canDisable['customers'] > 1) ? 2 : 1);
-            }
-
-            if (isset($canDisable['invoices'])) {
-                $text[] = '<b>' . $canDisable['invoices'] . '</b> ' . trans_choice('general.invoices', ($canDisable['invoices'] > 1) ? 2 : 1);
-            }
-
-            if (isset($canDisable['revenues'])) {
-                $text[] = '<b>' . $canDisable['revenues'] . '</b> ' . trans_choice('general.revenues', ($canDisable['revenues'] > 1) ? 2 : 1);
-            }
-
-            if (isset($canDisable['bills'])) {
-                $text[] = '<b>' . $canDisable['bills'] . '</b> ' . trans_choice('general.bills', ($canDisable['bills'] > 1) ? 2 : 1);
-            }
-
-            if (isset($canDisable['payments'])) {
-                $text[] = '<b>' . $canDisable['payments'] . '</b> ' . trans_choice('general.payments', ($canDisable['payments'] > 1) ? 2 : 1);
-            }
-
-            $message = trans('messages.warning.disabled', ['type' => trans_choice('general.currencies', 1), 'text' => implode(', ', $text)]);
+            $message = trans('messages.warning.disabled', ['name' => $currency->name, 'text' => implode(', ', $relationships)]);
 
             flash($message)->warning();
-        }
 
-        return redirect('settings/currencies');
+            return redirect('settings/currencies/' . $currency->id . '/edit');
+        }
     }
 
     /**
@@ -157,11 +140,30 @@ class Currencies extends Controller
      */
     public function destroy(Currency $currency)
     {
-        $currency->delete();
+        $relationships = $this->countRelationships($currency, [
+            'accounts' => 'accounts',
+            'customers' => 'customers',
+            'invoices' => 'invoices',
+            'revenues' => 'revenues',
+            'bills' => 'bills',
+            'payments' => 'payments',
+        ]);
 
-        $message = trans('messages.success.deleted', ['type' => trans_choice('general.currencies', 1)]);
+        if ($currency->code == setting('general.default_currency')) {
+            $relationships[] = strtolower(trans_choice('general.companies', 1));
+        }
 
-        flash($message)->success();
+        if (empty($relationships)) {
+            $currency->delete();
+
+            $message = trans('messages.success.deleted', ['type' => trans_choice('general.currencies', 1)]);
+
+            flash($message)->success();
+        } else {
+            $message = trans('messages.warning.deleted', ['name' => $currency->name, 'text' => implode(', ', $relationships)]);
+
+            flash($message)->warning();
+        }
 
         return redirect('settings/currencies');
     }
